+++ src/RenderSystem.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "Types.h"
 
 // Forward Declarations
 struct SDL_Renderer;
@@ -28,5 +29,5 @@ struct RenderSystem {
 void RenderSystem_Initialize(RenderSystem* renderSystem, ComponentBag* cBag, TileMap* tileMap);
 void RenderSystem_Render_xywh(SDL_Renderer* renderer, int x, int y, int w, int h, SDL_Rect* clip, Texture* texture);
 void RenderSystem_RenderCoord(SDL_Renderer* renderer, Rectangle* rect, SDL_Rect* clip, Texture* texture);
-void RenderSystem_Update(RenderSystem* renderSystem, SDL_Renderer* renderer, float delta);
+void RenderSystem_Update(RenderSystem* renderSystem, SDL_Renderer* renderer, uint32 delta);
 void RenderSystem_CompleteRender(SDL_Renderer* renderer);
